[FileBackend] Support "ignoreMissingSource" for copy and move operations.
* This lets callers use "copy if exist" semantics more easily and avoids extra stat
  queries to the backend (since the cache is cleared before doOperations()).
* Tweaked FileOp::fileSha1() to reduce backend stat requests as 404s are not cached.

+++ includes/filebackend/FSFileBackend.php
@@ -250,6 +250,13 @@ class FSFileBackend extends FileBackendStore {
 			return $status;
 		}
 
+		if ( !is_file( $source ) ) {
+			if ( empty( $params['ignoreMissingSource'] ) ) {
+				$status->fatal( 'backend-fail-copy', $params['src'] );
+			}
+			return $status; // do nothing; either OK or bad status
+		}
+
 		if ( file_exists( $dest ) ) {
 			$ok = unlink( $dest );
 			if ( !$ok ) {
@@ -308,6 +315,13 @@ class FSFileBackend extends FileBackendStore {
 		if ( $dest === null ) {
 			$status->fatal( 'backend-fail-invalidpath', $params['dst'] );
 			return $status;
+		}
+
+		if ( !is_file( $source ) ) {
+			if ( empty( $params['ignoreMissingSource'] ) ) {
+				$status->fatal( 'backend-fail-move', $params['src'] );
+			}
+			return $status; // do nothing; either OK or bad status
 		}
 
 		if ( file_exists( $dest ) ) {
